Moved EmailSchema and conf from manager.py to schemas.py
